# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Finances API",
    "version": "1.0.0",
    "description": "API para gerenciamento financeiro pessoal, com controle de contas, transações e categorias."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "paths": {
...
    "/categories": {
      "post": {
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "name",
                  "type"
                ],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "User ID"
                  },
                  "name": {
                    "type": "string",
                    "description": "Category name"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ],
                    "description": "Category type"
                  },
                  "color": {
                    "type": "string",
                    "description": "Category color (hex format)",
                    "example": "#FF5733"
                  },
                  "icon": {
                    "type": "string",
                    "description": "Category icon name",
                    "example": "utensils"
                  }
                },
                "example": {
                  "user_id": "456e7890-e89b-12d3-a456-426614174001",
                  "name": "Food",
                  "type": "expense",
                  "color": "#FF5733",
                  "icon": "utensils"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Category with this name already exists for this user"
          }
        }
      },
      "get": {
        "summary": "List all categories for a user",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "User ID"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            },
            "description": "Filter by category type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "400": {
            "description": "User ID not provided"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Get category by ID",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing category",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Category ID to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Category name"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ],
                    "description": "Category type"
                  },
                  "color": {
                    "type": "string",
                    "description": "Category color (hex format)"
                  },
                  "icon": {
                    "type": "string",
                    "description": "Category icon name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Category with this name already exists for this user"
          }
        }
      },
      "delete": {
        "summary": "Delete a category",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Category ID to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
...
  },
...
}